Make required changes
