+++ src/App.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Calculator from './components/Calculator';
+import './index.css';
+
+class App extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <Calculator />
+    );
+  }
+}
+
+export default App;
